+++ tests/conftest.py
@@ -1,9 +1,15 @@
 """Configuration for tests."""
 
+import json
 import os
 import subprocess
+import sys
 
 import _pytest
+import pytest
+
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.abspath(__file__), "..", "..")))
+import asconnect  # pylint: disable=wrong-import-order
 
 
 def pytest_configure(config: _pytest.config.Config) -> None:  # type: ignore
@@ -34,3 +40,38 @@ def pytest_configure(config: _pytest.config.Config) -> None:  # type: ignore
             continue
         name, value = line.split("=")
         os.environ[name] = value
+
+
+def get_test_data() -> tuple[str, str, str]:
+    """Get the test data.
+
+    :returns: The test data
+    """
+    tests_path = os.path.dirname(os.path.abspath(__file__))
+    test_data_path = os.path.join(tests_path, "test_data.json")
+    with open(test_data_path, "r", encoding="utf-8") as test_data_file:
+        all_test_data = json.load(test_data_file)
+
+    test_data = all_test_data["base"]
+
+    return test_data["key_id"], test_data["key"], test_data["issuer_id"]
+
+
+@pytest.fixture(scope="session")
+def client() -> asconnect.Client:
+    """Get the test client.
+
+    :returns: The test client
+    """
+    key_id, key, issuer_id = get_test_data()
+
+    return asconnect.Client(key_id=key_id, key_contents=key, issuer_id=issuer_id)
+
+
+@pytest.fixture(scope="session")
+def app_id() -> str:
+    """Get the test app ID.
+
+    :returns: The test app ID
+    """
+    return ""
